# app/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "Memory Game API",
        "description": "An API for managing and playing multiple Charades Against Humanity games."
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/game": {
            "get": {
                "tags": [
                    "Game"
                ],
                "summary": "Retrieves the current state of the memory game board.",
                "operationId": "game_get",
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "deck": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Card"
                                    }
                                },
                                "team1": {
                                    "$ref": "#/definitions/Team"
                                },
                                "team2": {
                                    "$ref": "#/definitions/Team"
                                },
                                "currentTurn": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/new_player": {
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Creates a new player object and assigns it to an existing game",
                "operationId": "game_new_player",
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "deprecated": false
            }
        },
        "/play_pass": {
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Gives current player the option to play or pass",
                "operationId": "game_play_pass",
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "playerId",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "deprecated": false
            }
        },
        "/actor_toggle": {
            "put": {
                "tags": [
                    "Game"
                ],
                "summary": "Changes player's actor status on/off",
                "operationId": "game_play_pass",
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "playerId",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "deprecated": false
            }
        },
        "/new": {
            "post": {
                "tags": [
                    "Game"
                ],
                "summary": "Initializes a new game board of the specified size (# of matches).",
                "operationId": "game_new",
                "produces": [
                    "application/json",
                    "text/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "team1Name",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "team2Name",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "deprecated": false
            }
        }
    },
    "definitions": {
        "Card": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "Player": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "teamId": {
                    "type": "string"
                },
                "score": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "Team": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "wonCards": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "id": {
                                "type": "string"
                            }
                        }
                    }
                },
                "lostCards": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "id": {
                                "type": "string"
                            }
                        }
                    }
                },
                "score": {
                    "type": "integer"
                },
                "players": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "teamId": {
                                "type": "string"
                            }
                        }
                    }
                },
                "actors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}